refactor(frontend): split settings views and normalize styles
