+++ src/components/WindowHeader.vue
@@ -19,19 +19,52 @@ function closeWindow() {
 <template>
   <header
     class="window-header"
+    h-40px w-full shrink-0 flex items-center
+    bg="$ui-page-background"
     data-tauri-drag-region
   >
-    <span class="window-title" data-tauri-drag-region>CodeNest</span>
+    <span
+      class="window-title"
+      relative left-30px lh-20px text-13px font-500
+      data-tauri-drag-region
+    >CodeNest</span>
 
-    <div class="window-controls">
-      <button class="window-control" type="button" :aria-label="t('app.window.minimize')" @click="minimizeWindow">
-        <span class="window-control-icon minimize" />
+    <div class="window-controls" h-full ml-auto flex items-center>
+      <button
+        class="window-control"
+        relative h-full w-46px border-0 bg-transparent
+        p-0
+        flex items-center justify-center
+        light:color="$gray-1" dark:color="$gray-11"
+        type="button"
+        :aria-label="t('app.window.minimize')"
+        @click="minimizeWindow"
+      >
+        <span class="window-control-icon minimize" relative block size-12px />
       </button>
-      <button class="window-control" type="button" :aria-label="t('app.window.maximize')" @click="toggleMaximizeWindow">
-        <span class="window-control-icon maximize" />
+      <button
+        class="window-control"
+        relative h-full w-46px border-0 bg-transparent
+        p-0
+        flex items-center justify-center
+        light:color="$gray-1" dark:color="$gray-11"
+        type="button"
+        :aria-label="t('app.window.maximize')"
+        @click="toggleMaximizeWindow"
+      >
+        <span class="window-control-icon maximize" relative block size-12px />
       </button>
-      <button class="window-control close" type="button" :aria-label="t('app.window.close')" @click="closeWindow">
-        <span class="window-control-icon close-icon" />
+      <button
+        class="window-control close"
+        relative h-full w-46px border-0 bg-transparent
+        p-0
+        flex items-center justify-center
+        light:color="$gray-1" dark:color="$gray-11"
+        type="button"
+        :aria-label="t('app.window.close')"
+        @click="closeWindow"
+      >
+        <span class="window-control-icon close-icon" relative block size-12px />
       </button>
     </div>
   </header>
@@ -39,27 +72,19 @@ function closeWindow() {
 
 <style lang="scss" scoped>
 .window-header {
-  @apply h-40px w-full shrink-0 flex items-center;
-  @apply bg-$ui-page-background;
   -webkit-app-region: drag;
   -webkit-user-select: none;
 }
 
 .window-controls {
-  @apply h-full ml-auto flex items-center;
   -webkit-app-region: no-drag;
 }
 
 .window-title {
-  @apply relative left-30px lh-20px text-13px font-500;
   font-family: var(--font-title);
 }
 
 .window-control {
-  @apply relative h-full w-46px border-0 bg-transparent p-0;
-  @apply flex items-center justify-center;
-  @apply color-$gray-1 dark:color-$gray-11;
-
   &:hover {
     @apply bg-$ui-hover-background;
   }
@@ -67,10 +92,6 @@ function closeWindow() {
   &.close:hover {
     @apply bg-$red-5 color-white;
   }
-}
-
-.window-control-icon {
-  @apply relative block size-12px;
 }
 
 .window-control-icon.minimize::before {
