+++ src/components/ui/UiDialog.vue
@@ -39,15 +39,34 @@ const contentStyle = computed(() => ({
     @update:open="value => emit('update:open', value)"
   >
     <DialogPortal>
-      <DialogOverlay class="ui-dialog-overlay" />
-      <DialogContent class="ui-dialog-content" :style="contentStyle">
-        <DialogTitle v-if="title || $slots.title" class="ui-dialog-title">
+      <DialogOverlay
+        class="ui-dialog-overlay"
+        fixed left-0 right-0 bottom-0 z-40
+        bg="black/30" dark:bg="black/50"
+      />
+      <DialogContent
+        class="ui-dialog-content left-1/2 -translate-x-1/2 -translate-y-1/2"
+        fixed z-50 max-w="[calc(100vw-32px)]"
+        rounded-6px border p-20px outline-none
+        bg="$ui-surface-background" color="$ui-foreground"
+        :style="contentStyle"
+      >
+        <DialogTitle
+          v-if="title || $slots.title"
+          m-0 text-15px font-650 lh-20px
+          class="ui-dialog-title"
+        >
           <slot name="title">
             {{ title }}
           </slot>
         </DialogTitle>
 
-        <DialogDescription v-if="description || $slots.description" class="ui-dialog-description">
+        <DialogDescription
+          v-if="description || $slots.description"
+          mt-8px mb-0 text-13px lh-18px
+          light:color="$gray-5" dark:color="$gray-9"
+          class="ui-dialog-description"
+        >
           <slot name="description">
             {{ description }}
           </slot>
@@ -55,7 +74,12 @@ const contentStyle = computed(() => ({
 
         <slot />
 
-        <div v-if="$slots.actions" class="ui-dialog-actions" :data-layout="actionsLayout">
+        <div
+          v-if="$slots.actions"
+          class="ui-dialog-actions"
+          mt-20px flex flex-wrap gap-8px
+          :data-layout="actionsLayout"
+        >
           <slot name="actions" />
         </div>
       </DialogContent>
@@ -65,15 +89,11 @@ const contentStyle = computed(() => ({
 
 <style lang="scss" scoped>
 .ui-dialog-overlay {
-  @apply fixed left-0 right-0 bottom-0 z-40;
-  @apply bg-black/30 dark:bg-black/50;
   top: var(--window-titlebar-height, 40px);
 }
 
 .ui-dialog-content {
-  @apply fixed left-1/2 z-50 max-w-[calc(100vw-32px)];
-  @apply -translate-x-1/2 -translate-y-1/2 rounded-6px border p-20px outline-none;
-  @apply border-$ui-border bg-$ui-surface-background color-$ui-foreground;
+  border-color: var(--ui-border);
   border-style: solid;
   box-shadow: var(--shadow-dialog);
   top: calc(var(--window-titlebar-height, 40px) + 50%);
@@ -88,18 +108,14 @@ const contentStyle = computed(() => ({
 }
 
 .ui-dialog-title {
-  @apply m-0 text-15px font-650 lh-20px;
   overflow-wrap: anywhere;
 }
 
 .ui-dialog-description {
-  @apply mt-8px mb-0 text-13px lh-18px light:color-$gray-5 dark:color-$gray-9;
   overflow-wrap: anywhere;
 }
 
 .ui-dialog-actions {
-  @apply mt-20px flex flex-wrap gap-8px;
-
   &[data-layout="reverse"] {
     @apply flex-row-reverse;
   }
