+++ src/components/ui/UiScrollArea.vue
@@ -17,35 +17,42 @@ withDefaults(defineProps<{
 </script>
 
 <template>
-  <ScrollAreaRoot class="ui-scroll-root" :type="type">
-    <ScrollAreaViewport class="ui-scroll-viewport">
+  <ScrollAreaRoot class="ui-scroll-root" min-h-0 min-w-0 overflow-hidden :type="type">
+    <ScrollAreaViewport class="ui-scroll-viewport" size-full min-h-0 min-w-0>
       <slot />
     </ScrollAreaViewport>
 
-    <ScrollAreaScrollbar class="ui-scrollbar" orientation="vertical">
-      <ScrollAreaThumb class="ui-scroll-thumb" />
+    <ScrollAreaScrollbar
+      class="ui-scrollbar" z-10 flex touch-none select-none
+      p-1px bg-transparent transition-colors orientation="vertical"
+    >
+      <ScrollAreaThumb
+        class="ui-scroll-thumb" relative flex-1 rounded-full light:bg="$gray-9/45"
+        dark:bg="$gray-6/50"
+      />
     </ScrollAreaScrollbar>
-    <ScrollAreaScrollbar v-if="horizontal" class="ui-scrollbar horizontal" orientation="horizontal">
-      <ScrollAreaThumb class="ui-scroll-thumb" />
+    <ScrollAreaScrollbar
+      v-if="horizontal"
+      class="ui-scrollbar horizontal"
+      z-10 flex touch-none select-none p-1px
+      bg-transparent transition-colors
+      orientation="horizontal"
+    >
+      <ScrollAreaThumb
+        class="ui-scroll-thumb" relative flex-1 rounded-full light:bg="$gray-9/45"
+        dark:bg="$gray-6/50"
+      />
     </ScrollAreaScrollbar>
-    <ScrollAreaCorner v-if="horizontal" class="ui-scroll-corner" />
+    <ScrollAreaCorner v-if="horizontal" bg-transparent />
   </ScrollAreaRoot>
 </template>
 
 <style lang="scss">
-.ui-scroll-root {
-  @apply min-h-0 min-w-0 overflow-hidden;
-}
-
 .ui-scroll-viewport {
-  @apply size-full min-h-0 min-w-0;
   scrollbar-gutter: stable;
 }
 
 .ui-scrollbar {
-  @apply z-10 flex touch-none select-none p-1px;
-  @apply bg-transparent transition-colors;
-
   &[data-orientation="vertical"] {
     @apply w-6px;
   }
@@ -56,15 +63,8 @@ withDefaults(defineProps<{
 }
 
 .ui-scroll-thumb {
-  @apply relative flex-1 rounded-full;
-  @apply light:bg-$gray-9/45 dark:bg-$gray-6/50;
-
   &:hover {
     @apply light:bg-$gray-7/60 dark:bg-$gray-7/65;
   }
-}
-
-.ui-scroll-corner {
-  @apply bg-transparent;
 }
 </style>
