+++ src/components/ui/UiSegmentedControl.vue
@@ -35,11 +35,22 @@ function updateValue(value: unknown) {
     :model-value="modelValue"
     @update:model-value="updateValue"
   >
-    <div class="ui-segmented-control" :aria-label="ariaLabel">
+    <div
+      class="ui-segmented-control"
+      h-27px rounded-5px p-2px flex items-center
+      dark:bg="$ui-control-background"
+      light:bg="[color-mix(in_srgb,var(--ui-switch-background)_46%,var(--ui-surface-background))]"
+      :aria-label="ariaLabel"
+    >
       <ToggleGroupItem
         v-for="option in options"
         :key="option.value"
         class="ui-segmented-button"
+        h-24px border-0 rounded-4px px-8px bg-transparent
+        inline-flex items-center justify-center gap-5px whitespace-nowrap
+        text-12px cursor-pointer outline-none
+        light:color="$gray-5" dark:color="$gray-9"
+        transition duration-120 ease-out
         :class="{ active: modelValue === option.value }"
         :value="option.value"
         :title="option.tooltip"
@@ -54,18 +65,7 @@ function updateValue(value: unknown) {
 </template>
 
 <style lang="scss">
-.ui-segmented-control {
-  @apply h-27px rounded-5px p-2px flex items-center;
-  @apply dark:bg-$ui-control-background;
-  @apply light:[background:color-mix(in_srgb,var(--ui-switch-background)_46%,var(--ui-surface-background))];
-}
-
 .ui-segmented-button {
-  @apply h-24px border-0 rounded-4px px-8px bg-transparent;
-  @apply inline-flex items-center justify-center gap-5px whitespace-nowrap;
-  @apply text-12px cursor-pointer outline-none light:color-$gray-5 dark:color-$gray-9;
-  @apply transition duration-120 ease-out;
-
   &:hover:not(.active):not(:disabled) {
     @apply color-$ui-foreground;
     background-color: color-mix(in srgb, var(--ui-hover-background), var(--ui-foreground) 5%);
