+++ src/constants/storageKeys.ts
@@ -1,0 +1,1 @@
+export const HOME_LAYOUT_STORAGE_KEY = 'codenest:home-layout'
