+++ src/views/SettingsView/components/SettingsRow.vue
@@ -18,19 +18,29 @@ withDefaults(defineProps<{
 </script>
 
 <template>
-  <div class="setting-row" :class="{ 'align-start': align === 'start' }">
-    <div class="setting-copy">
-      <strong>{{ title }}</strong>
-      <span v-if="description" :class="descriptionTone">
+  <div
+    class="setting-row"
+    min-h-42px min-w-0 flex items-center justify-between
+    gap-14px
+    :class="{ 'align-start': align === 'start' }"
+  >
+    <div min-w-0 flex-1 flex flex-col gap-3px>
+      <strong text-13px font-620>{{ title }}</strong>
+      <span v-if="description" text-12px light:color="$gray-6" dark:color="$gray-8" :class="descriptionTone">
         {{ description }}
       </span>
     </div>
 
-    <div class="setting-content">
+    <div
+      class="setting-content" min-w-0 max-w-full flex flex-col
+      items-end gap-6px
+    >
       <slot />
       <p
         v-if="reserveStatus || status || $slots.status"
         class="setting-status"
+        m-0 min-h-16px max-w-520px text-right text-12px
+        lh-16px
         :class="[statusTone, { empty: !status && !$slots.status }]"
         :aria-hidden="!status && !$slots.status"
       >
@@ -43,10 +53,6 @@ withDefaults(defineProps<{
 </template>
 
 <style lang="scss" scoped>
-.setting-row {
-  @apply min-h-42px min-w-0 flex items-center justify-between gap-14px;
-}
-
 .setting-row.align-start {
   @apply items-start;
 
@@ -55,30 +61,16 @@ withDefaults(defineProps<{
   }
 }
 
-.setting-copy {
-  @apply min-w-0 flex-1 flex flex-col gap-3px;
-
-  strong {
-    @apply text-13px font-620;
-    overflow-wrap: anywhere;
-  }
-
-  span {
-    @apply text-12px light:color-$gray-6 dark:color-$gray-8;
-    overflow-wrap: anywhere;
-  }
-
-  span.danger {
-    @apply color-$red-5;
-  }
+strong,
+span {
+  overflow-wrap: anywhere;
 }
 
-.setting-content {
-  @apply min-w-0 max-w-full flex flex-col items-end gap-6px;
+span.danger {
+  @apply color-$red-5;
 }
 
 .setting-status {
-  @apply m-0 min-h-16px max-w-520px text-right text-12px lh-16px;
   overflow-wrap: anywhere;
 
   &.empty {
